+++ SwiftMessages/Animator.swift
@@ -67,17 +67,12 @@ public protocol Animator: class {
 
     func hide(context: AnimationContext, completion: @escaping AnimationCompletion)
 
-    /// The show animation duration. Return `nil` (the default) if animation has undefined duration,
-    /// such as if using `UIDynamicAnimator`. This value is utilized by `SwiftMessagesSegue`.
-    var showDuration: TimeInterval? { get }
+    /// The show animation duration. If the animation duration is unknown, such as if using `UIDynamnicAnimator`,
+    /// then profide an estimate. This value is used by `SwiftMessagesSegue`.
+    var showDuration: TimeInterval { get }
 
-    /// The hide animation duration. Return `nil` (the default) if animation has undefined duration,
-    /// such as if using `UIDynamicAnimator`. This value is utilized by `SwiftMessagesSegue`.
-    var hideDuration: TimeInterval? { get }
-}
-
-public extension Animator {
-    var showDuration: TimeInterval? { return nil }
-    var hideDuration: TimeInterval? { return nil }
+    /// The hide animation duration. If the animation duration is unknown, such as if using `UIDynamnicAnimator`,
+    /// then profide an estimate. This value is used by `SwiftMessagesSegue`.
+    var hideDuration: TimeInterval { get }
 }